Adjust API and efficiency
This includes the parallel moves calculation. From this commit you can choose if you want a sequantial calculation of a parallel one with one you can be ~70% more efficient!
