+++ .golangci.yml
@@ -3,9 +3,6 @@ linters-settings:
     min-complexity: 15
   lll:
     line-length: 100
-  errcheck:
-    check-type-assertions: true
-    check-blank: true
   revive:
     ignore-generated-header: true
     severity: warning
