+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [#16](https://github.com/RchrdHndrcks/gochess/pull/16) Adjust API and efficiency.
+
 ## [1.1.0] - 2025-04-23
 
 ### Changed
