+++ README.md
@@ -89,7 +89,7 @@ Width() int
 Square(c Coordinate) (int8, error)
 MakeMove(origin, target Coordinate) error
 SetSquare(c Coordinate, p int8) error
-isValidCoordinate(c Coordinate) bool
+Clone() *Board
 ```
 
 ### Coordinates
